+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Initial strategy: farthest point sampling
+
 ### Fixed
 - UCB now usable as acquisition function, hard-set beta parameter to 1.0.